multiple hash algorithms using system's crypt in pacrypt
This patch adds another $CONF['encrypt'] type, 'php_crypt', which uses the PHP crypt() function, which in turn uses the system's crypt() (much like the current 'system' type in postfixadmin-2.3.6)

The difference to the current 'system' type is that you can choose which hash type you want to use for new passwords. (currently DES, MD5, Blowfish, SHA256, SHA512; defaults to MD5)

Example:
$CONF['encrypt'] = 'php_crypt:SHA512';

The current 'system' type only generates MD5 hashes with 2-character salt.

I did not have 'dovecotpw' on my system so that was not an option.

It does not interfere with existing 'encrypt' types, but could be incorporated into the 'system' type.
Tested on CentOS 5, 6 and 7, Dovecot 2 and Postfix.