add to vacation.pl
function get_accountname

get the variable name from the table mailbox and use it as accountname
if accountname_check is true the accountname is placed in front of the sender's email " Accountname <original@recipient.domain>" and this will be sent as email name and address of the sender.
if the account is not filled with a string then we will be checking if friendly_name is filled with a string and this will be placed in front of the sender's email "Friendlyname <original@recipient.domain>" and this will be sent as the email name and address of the sender.
if not, only the sender's email address will be sent.

function replace_sting together with replace_a_string

replaces in the body text to be sent the texts describing $replace_from and $replace_until with the values ​​of activerom and activeuntil respectively from the vacation table in the date format specified at $date_format
the date format capabilities are described at https://www.tutorialspoint.com/perl/perl_date_time.htm. Variable has been chosen in the config part of vacation.pl so that admin can adjust the search texts for a country version.
in the languages/??.lang file one can see the $PALANG['UsersVacation_body_text'] with the desired replacement texts in $Replace_From and $Replace_Until.

Added library Time::Piece for data format