Ignore CSV labels with a whitespace
Currently, exporting the alias list to CSV, the first column label is only a whitespace. For some reason the contents of this column consists of a long concatenation of "nbsp;"s.
This fix ignores the whitespace column label and gets rid of the "nbsp;nbsp;......" contents of this column.
This was successfully tested on a running installation of postfixadmin 3.2 on a Debian 11 system.